Improve subrouter behavior plus a few other fixes
Improve subrouter behavior by allowing the subrouter to resolve the URI
relative to the URI matched from the regex that picked the subrouter.
This allows all subrouter URIs to be relative to the router's new base
path. This does not alter the base path property since it's needed for
getURI() to work properly.

Fix handling of query parameter array in getURI() to check to empty
array.

Fix default routes for HEAD and OPTIONS. The user may now specify HEAD
and OPTION routes that will be given precedence over the default routes.