Add method for adding set of routes in table
Adds a method to the Router class that allows users to specify routes in
bulk instead of one per method call.

Also, we set the statusCode to 404 by default when the not found handler
is applied. This will only propogate to the HTTP response code if the
user employs the Router's flush() method.