Avoid result handling after calling a subrouter
If a subrouter is called, then the parent router should *not* handle the
result. (The result is actually null since Router::routeImpl() does not
return a value.) The fix added a unique class RouterClosure that wraps a
Closure to call the subrouter. This class can be identified in the case
when we are calling a subrouter.

This also fixes a bug where the callable check (i.e. via return type
hint) would fail for an array containing a private method in a derived
router class.